odoo_function/requirements.txt
annotated-types==0.6.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
deprecation==2.1.0
dnspython==2.4.2
email-validator==2.1.0.post1
flake8==6.1.0
Flask==3.0.0
functions-framework==3.*
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-secret-manager==2.17.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
OdooRPC==0.10.1
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==3.1.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
typing_extensions==4.8.0
urllib3==2.1.0
watchdog==3.0.0
Werkzeug==3.0.1
google-cloud-logging==3.*